+++ CMakeLists.txt
@@ -586,13 +586,6 @@ if(BUILD_TESTS)
   )
 
   add_e2e_test(
-    NAME content_types_test
-    PYTHON_SCRIPT ${CMAKE_SOURCE_DIR}/tests/js-content-types/content_types.py
-    CONSENSUS cft
-    ADDITIONAL_ARGS --js-app-bundle ${CMAKE_SOURCE_DIR}/tests/js-content-types
-  )
-
-  add_e2e_test(
     NAME launch_host_process_test
     PYTHON_SCRIPT
       ${CMAKE_SOURCE_DIR}/tests/js-launch-host-process/host_process.py
