+++ tests/js-custom-authorization/custom_authorization.py
@@ -213,6 +213,72 @@ def run_authn(args):
         network = test_role_based_access(network, args)
 
 
+@reqs.description("Test content types")
+def test_content_types(network, args):
+    primary, _ = network.find_nodes()
+
+    with primary.client("user0") as c:
+        r = c.post("/app/text", body="text")
+        assert r.status_code == http.HTTPStatus.OK, r.status_code
+        assert r.headers["content-type"] == "text/plain"
+        assert r.body.text() == "text"
+
+        r = c.post("/app/json", body={"foo": "bar"})
+        assert r.status_code == http.HTTPStatus.OK, r.status_code
+        assert r.headers["content-type"] == "application/json"
+        assert r.body.json() == {"foo": "bar"}
+
+        r = c.post("/app/binary", body=b"\x00" * 42)
+        assert r.status_code == http.HTTPStatus.OK, r.status_code
+        assert r.headers["content-type"] == "application/octet-stream"
+        assert r.body.data() == b"\x00" * 42, r.body
+
+        r = c.post("/app/custom", body="text", headers={"content-type": "foo/bar"})
+        assert r.status_code == http.HTTPStatus.OK, r.status_code
+        assert r.headers["content-type"] == "text/plain"
+        assert r.body.text() == "text"
+
+    return network
+
+
+@reqs.description("Test unknown path")
+def test_unknown_path(network, args):
+    primary, _ = network.find_nodes()
+
+    with primary.client("user0") as c:
+        r = c.get("/app/not/a/real/path")
+        assert r.status_code == http.HTTPStatus.NOT_FOUND, r.status_code
+        r = c.post("/app/not/a/real/path")
+        assert r.status_code == http.HTTPStatus.NOT_FOUND, r.status_code
+        r = c.delete("/app/not/a/real/path")
+        assert r.status_code == http.HTTPStatus.NOT_FOUND, r.status_code
+
+        r = c.post("/app/unknown")
+        assert r.status_code == http.HTTPStatus.NOT_FOUND, r.status_code
+
+    with primary.client() as c:
+        r = c.get("/app/not/a/real/path")
+        assert r.status_code == http.HTTPStatus.NOT_FOUND, r.status_code
+        r = c.post("/app/not/a/real/path")
+        assert r.status_code == http.HTTPStatus.NOT_FOUND, r.status_code
+        r = c.delete("/app/not/a/real/path")
+        assert r.status_code == http.HTTPStatus.NOT_FOUND, r.status_code
+
+        r = c.post("/app/unknown")
+        assert r.status_code == http.HTTPStatus.NOT_FOUND, r.status_code
+
+    return network
+
+
+def run_content_types(args):
+    with infra.network.network(
+        args.nodes, args.binary_dir, args.debug_nodes, args.perf_nodes, pdb=args.pdb
+    ) as network:
+        network.start_and_join(args)
+        network = test_content_types(network, args)
+        network = test_unknown_path(network, args)
+
+
 if __name__ == "__main__":
     cr = ConcurrentRunner()
 
@@ -220,7 +286,7 @@ if __name__ == "__main__":
         "authz",
         run,
         package="libjs_generic",
-        nodes=infra.e2e_args.max_nodes(cr.args, f=0),
+        nodes=infra.e2e_args.nodes(cr.args, 1),
         js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-custom-authorization"),
     )
 
@@ -228,7 +294,7 @@ if __name__ == "__main__":
         "limits",
         run_limits,
         package="libjs_generic",
-        nodes=infra.e2e_args.max_nodes(cr.args, f=0),
+        nodes=infra.e2e_args.nodes(cr.args, 1),
         js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-limits"),
     )
 
@@ -236,10 +302,18 @@ if __name__ == "__main__":
         "authn",
         run_authn,
         package="libjs_generic",
-        nodes=infra.e2e_args.max_nodes(cr.args, f=0),
+        nodes=infra.e2e_args.nodes(cr.args, 1),
         js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-authentication"),
         initial_user_count=4,
         initial_member_count=2,
+    )
+
+    cr.add(
+        "content_types",
+        run_content_types,
+        package="libjs_generic",
+        nodes=infra.e2e_args.nodes(cr.args, 1),
+        js_app_bundle=os.path.join(cr.args.js_app_bundle, "js-content-types"),
     )
 
     cr.run()
